requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
psycopg2-binary==2.9.6
sqlalchemy==2.0.9
redis==4.5.4
elasticsearch==8.7.0
python-dotenv==1.0.0
pydantic==1.10.7
